Out of working hours changes
When out of working hours, an icon is now displayed instead of text
(with a tooltip to give more info).
Unfortunately, to do this, I had to completely dismantle the
'issue_fields_rows' table structure and convert it to the raw html. This
was necessary, because otherwise the icon and tooltip options get added
to both the title and the text of the resulting row.

Also, this icon does not appear when in breach.
The icon itself is also temporary (an hourglass). Will get a proper one
when I hear back from Stig.